+++ packages/plugin-timeclock-ui/src/components/schedule/ScheduleList.tsx
@@ -1,28 +1,15 @@
 import Button from '@erxes/ui/src/components/Button';
-import { router, __ } from '@erxes/ui/src/utils';
+import { __ } from '@erxes/ui/src/utils';
 import React, { useState } from 'react';
 import ModalTrigger from '@erxes/ui/src/components/ModalTrigger';
 import Wrapper from '@erxes/ui/src/layout/components/Wrapper';
 import Table from '@erxes/ui/src/components/table';
-import SelectTeamMembers from '@erxes/ui/src/team/containers/SelectTeamMembers';
 import NameCard from '@erxes/ui/src/components/nameCard/NameCard';
-import DateRange from '../datepicker/DateRange';
-import dayjs from 'dayjs';
-import DatePicker from '../datepicker/DateTimePicker';
-import { ISchedule } from '../../types';
-import ScheduleConfig from './ScheduleDayToggleConfig';
-import Select from 'react-select-plus';
-import SelectDepartments from '@erxes/ui-settings/src/departments/containers/SelectDepartments';
-import { CustomRow, Input } from '../../styles';
+import { CustomRow } from '../../styles';
 
-import FormGroup from '@erxes/ui/src/components/form/Group';
-import ControlLabel from '@erxes/ui/src/components/form/Label';
-import { Row } from '../../styles';
 import { IBranch } from '@erxes/ui/src/team/types';
-import { CustomRangeContainer } from '../../styles';
-import DateControl from '@erxes/ui/src/components/form/DateControl';
-import Datetime from '@nateradebaugh/react-datetime';
 import Tip from '@erxes/ui/src/components/Tip';
+import ScheduleForm from './ScheduleForm';
 
 type Props = {
   scheduleOfMembers: any;
@@ -39,567 +26,50 @@ type Props = {
 
 function ScheduleList(props: Props) {
   const {
-    queryParams,
-    submitRequest,
-    submitShift,
-    history,
     scheduleOfMembers,
     solveSchedule,
     solveShift,
     getActionBar,
-    removeScheduleShifts,
-    branchesList
+    removeScheduleShifts
   } = props;
 
-  const [dateKeyCounter, setKeyCounter] = useState('');
-  const [userIds, setUserIds] = useState(['']);
-
   const trigger = (
-    <Button id="timeClockButton2" btnStyle="success" icon="plus-circle">
+    <Button btnStyle="success" icon="plus-circle">
       Create Request
     </Button>
   );
 
   const adminTrigger = (
-    <Button id="timeClockButton2" btnStyle="success" icon="plus-circle">
+    <Button btnStyle="success" icon="plus-circle">
       Create Request - Admin
     </Button>
   );
 
   const adminConfigTrigger = (
-    <Button id="timeClockButton2" btnStyle="primary" icon="plus-circle">
+    <Button btnStyle="primary" icon="plus-circle">
       Schedule Configuration - Admin
     </Button>
   );
 
-  const timeFormat = 'HH:mm';
-  const [dateRangeStart, setDateStart] = useState(new Date());
-  const [dateRangeEnd, setDateEnd] = useState(new Date());
-  const [scheduleDates, setScheduleDates] = useState<ISchedule>({});
-  const [contentType, setContentType] = useState('');
-  const [selectedDeptIds, setDepartments] = useState(['']);
-  const [selectedBranchIds, setBranches] = useState(['']);
-  const [configDays, setConfigDays] = useState<ISchedule>({
-    Monday: { display: true },
-    Tuesday: { display: true },
-    Wednesday: { display: true },
-    Thursday: { display: true },
-    Friday: { display: true },
-    Saturday: { display: true },
-    Sunday: { display: true }
-  });
-
-  const renderBranchOptions = (branches: any[]) => {
-    return branches.map(branch => ({
-      value: branch._id,
-      label: branch.title
-    }));
-  };
-
-  const onBranchSelect = selectedBranch => {
-    setBranches(selectedBranch);
-  };
-
-  const onDepartmentSelect = dept => {
-    setDepartments(dept);
-  };
-
-  const onRemoveDate = day_key => {
-    delete scheduleDates[day_key];
-    setScheduleDates({
-      ...scheduleDates
-    });
-    setKeyCounter(Object.keys(scheduleDates).at(-1) || '');
-  };
-
-  const onDateChange = (day_key, selectedDate) => {
-    const newShift = scheduleDates[day_key];
-
-    const oldShiftEnd = newShift.shiftEnd;
-    const oldShiftStart = newShift.shiftStart;
-
-    const newShiftStart = dayjs(
-      selectedDate.toLocaleDateString() +
-        ' ' +
-        dayjs(oldShiftStart).format(timeFormat)
-    ).toDate();
-    const newShiftEnd = dayjs(
-      selectedDate.toLocaleDateString() +
-        ' ' +
-        dayjs(oldShiftEnd).format(timeFormat)
-    ).toDate();
-
-    newShift.shiftDate = selectedDate;
-    newShift.shiftStart = newShiftStart;
-    newShift.shiftEnd = newShiftEnd;
-
-    const newScheduleDates = { ...scheduleDates, [day_key]: newShift };
-    setScheduleDates(newScheduleDates);
-  };
-
-  const compareStartAndEndTime = (day_key, newShiftStart, newShiftEnd) => {
-    const currShift = scheduleDates[day_key];
-    const currShiftDate = currShift.shiftDate?.toLocaleDateString();
-    const currShiftEnd = newShiftEnd ? newShiftEnd : currShift.shiftEnd;
-    const currShiftStart = newShiftStart ? newShiftStart : currShift.shiftStart;
-
-    let overnightShift = false;
-    let correctShiftEnd;
-
-    if (
-      dayjs(
-        currShiftDate + ' ' + dayjs(currShiftEnd).format(timeFormat)
-      ).format(timeFormat) < dayjs(currShiftStart).format(timeFormat)
-    ) {
-      correctShiftEnd =
-        dayjs(currShiftDate)
-          .add(1, 'day')
-          .toDate()
-          .toLocaleDateString() +
-        ' ' +
-        dayjs(currShiftEnd).format(timeFormat);
-
-      overnightShift = true;
-    } else {
-      correctShiftEnd = dayjs(
-        currShiftDate + ' ' + dayjs(currShiftEnd).format(timeFormat)
-      ).toDate();
-    }
-
-    const correctShiftStart = dayjs(
-      currShiftDate + ' ' + dayjs(currShiftStart).format(timeFormat)
-    ).toDate();
-
-    return [correctShiftStart, correctShiftEnd, overnightShift];
-  };
-
-  const onStartTimeChange = (day_key, time) => {
-    const newShift = scheduleDates[day_key];
-    const [
-      getCorrectStartTime,
-      getCorrectEndTime,
-      overnight
-    ] = compareStartAndEndTime(day_key, time, null);
-
-    newShift.shiftStart = getCorrectStartTime;
-    newShift.overnightShift = overnight;
-    newShift.shiftEnd = getCorrectEndTime;
-
-    const newScheduleDates = { ...scheduleDates, [day_key]: newShift };
-    setScheduleDates(newScheduleDates);
-  };
-
-  const onEndTimeChange = (day_key, time) => {
-    const newShift = scheduleDates[day_key];
-    const [
-      getCorrectStartTime,
-      getCorrectEndTime,
-      overnight
-    ] = compareStartAndEndTime(day_key, null, time);
-
-    newShift.shiftStart = getCorrectStartTime;
-    newShift.overnightShift = overnight;
-    newShift.shiftEnd = getCorrectEndTime;
-
-    const newScheduleDates = { ...scheduleDates, [day_key]: newShift };
-    setScheduleDates(newScheduleDates);
-  };
-
-  const pickSubset = Object.values(scheduleDates).map(shift => {
-    return { shiftStart: shift.shiftStart, shiftEnd: shift.shiftEnd };
-  });
-
-  const onSubmitClick = closeModal => {
-    submitRequest(userIds, pickSubset);
-    closeModal();
-  };
-  const onAdminSubmitClick = closeModal => {
-    submitShift(userIds, pickSubset);
-    closeModal();
-  };
-  const onUserSelect = users => {
-    setUserIds(users);
-  };
-
-  const clearDays = () => {
-    setScheduleDates({});
-  };
-
-  const addDay = () => {
-    // sort array of dates, in order to get the latest day
-    let dates_arr = Object.values(scheduleDates).map(shift => shift.shiftDate);
-    dates_arr = dates_arr.sort((a, b) => b.getTime() - a.getTime());
-
-    const dates = scheduleDates;
-    const getLatestDayKey = dates_arr
-      ? dayjs(dates_arr[0])
-          .add(1, 'day')
-          .toDate()
-          .toLocaleDateString()
-      : new Date().toLocaleDateString();
-
-    dates[getLatestDayKey] = {
-      shiftDate: new Date(getLatestDayKey),
-      shiftStart: new Date(getLatestDayKey + ' 08:30:00'),
-      shiftEnd: new Date(getLatestDayKey + ' 17:00:00')
-    };
-
-    setScheduleDates({
-      ...dates
-    });
-    setKeyCounter(getLatestDayKey);
-  };
-
-  const renderWeekDays = () => {
-    return (
-      <>
-        {Object.keys(scheduleDates).map(date_key => {
-          return (
-            <DatePicker
-              key={date_key}
-              startDate={scheduleDates[date_key].shiftDate}
-              startTime_value={scheduleDates[date_key].shiftStart}
-              endTime_value={scheduleDates[date_key].shiftEnd}
-              overnightShift={scheduleDates[date_key].overnightShift}
-              curr_day_key={date_key}
-              changeDate={onDateChange}
-              removeDate={onRemoveDate}
-              changeEndTime={onEndTimeChange}
-              changeStartTime={onStartTimeChange}
-            />
-          );
-        })}
-      </>
-    );
-  };
-
   const modalContent = ({ closeModal }) => (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: '10px' }}>
-      <SelectTeamMembers
-        queryParams={queryParams}
-        label={'Team member'}
-        onSelect={onUserSelect}
-        multi={false}
-        name="userId"
-      />
-      {renderWeekDays()}
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'row',
-          justifyContent: 'center'
-        }}
-      >
-        <Button style={{ marginTop: 10 }} onClick={clearDays}>
-          Clear
-        </Button>
-        <Button style={{ marginTop: 10 }} onClick={addDay}>
-          Add day
-        </Button>
-        <Button
-          style={{ marginTop: 10 }}
-          btnStyle="success"
-          onClick={() => onSubmitClick(closeModal)}
-        >
-          {'Submit'}
-        </Button>
-      </div>
-    </div>
+    <ScheduleForm modalContentType={''} closeModal={closeModal} {...props} />
   );
 
   const adminModalContent = ({ closeModal }) => (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: '10px' }}>
-      <SelectTeamMembers
-        queryParams={queryParams}
-        label={'Team member'}
-        onSelect={onUserSelect}
-        name="userId"
+    <ScheduleForm
+      modalContentType={'admin'}
+      closeModal={closeModal}
+      {...props}
+    />
+  );
+
+  const adminConfigContent = ({ closeModal }) => {
+    return (
+      <ScheduleForm
+        modalContentType={'adminConfig'}
+        closeModal={closeModal}
+        {...props}
       />
-      {renderWeekDays()}
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'row',
-          justifyContent: 'center'
-        }}
-      >
-        <Button style={{ marginTop: 10 }} onClick={clearDays}>
-          Clear
-        </Button>
-        <Button style={{ marginTop: 10 }} onClick={addDay}>
-          Add day
-        </Button>
-        <Button
-          btnStyle="success"
-          style={{ marginTop: 10 }}
-          onClick={() => onAdminSubmitClick(closeModal)}
-        >
-          {'Submit'}
-        </Button>
-      </div>
-    </div>
-  );
-
-  const toggleWeekDays = dayKey => {
-    const oldConfigBoolean = {
-      ...configDays[dayKey],
-      display: !configDays[dayKey].display
-    };
-    const newConfigDays = { ...configDays, [dayKey]: oldConfigBoolean };
-    setConfigDays(newConfigDays);
-  };
-
-  const renderWeekendSettings = () => {
-    return (
-      <>
-        {Object.keys(configDays).map(weekDay => (
-          <ScheduleConfig
-            key={weekDay}
-            weekDay={weekDay}
-            toggleWeekDays={toggleWeekDays}
-          />
-        ))}
-      </>
-    );
-  };
-
-  const renderConfigDays = () => {
-    return Object.keys(configDays).map(configDay => {
-      return (
-        <div
-          key={configDay}
-          style={{
-            display: 'flex',
-            flexDirection: 'row',
-            justifyContent: 'space-between'
-          }}
-        >
-          {configDays[configDay].display && (
-            <>
-              {configDay}
-              <div style={{ display: 'flex', flexDirection: 'row' }}>
-                <Datetime dateFormat={false} input={true} timeFormat="HH:mm" />
-                <Datetime
-                  defaultValue={new Date()}
-                  dateFormat={false}
-                  timeFormat="HH:mm"
-                />
-              </div>
-            </>
-          )}
-        </div>
-      );
-    });
-  };
-
-  const onDateRangeStartChange = (newStart: Date) => {
-    setDateStart(newStart);
-  };
-
-  const onDateRangeEndChange = (newEnd: Date) => {
-    setDateEnd(newEnd);
-  };
-
-  const onSaveDateRange = () => {
-    const format = 'YYYY-MM-DD HH:mm';
-    const formattedStartDate = dayjs(dateRangeStart).format(format);
-    const formattedEndDate = dayjs(dateRangeEnd).format(format);
-
-    router.setParams(history, {
-      startDate: formattedStartDate,
-      endDate: formattedEndDate
-    });
-
-    const totalDatesArray: string[] = [];
-
-    let temp = dayjs(dateRangeStart);
-
-    const endRange = dayjs(dateRangeEnd);
-
-    while (temp <= endRange) {
-      totalDatesArray.push(temp.toDate().toDateString());
-      temp = temp.add(1, 'day');
-    }
-
-    const newDatesByRange: ISchedule = scheduleDates;
-
-    for (const eachDay of totalDatesArray) {
-      newDatesByRange[eachDay] = {
-        shiftStart: new Date(eachDay),
-        shiftEnd: new Date(eachDay)
-      };
-
-      setKeyCounter(eachDay);
-    }
-
-    const difference = Object.keys(newDatesByRange).filter(
-      x => !totalDatesArray.includes(x)
-    );
-
-    for (const removeKey of difference) {
-      delete newDatesByRange[removeKey];
-    }
-
-    setScheduleDates(newDatesByRange);
-  };
-
-  const adminConfigByDate = () => {
-    return (
-      <div style={{ display: 'flex', flexDirection: 'column', gap: '20px' }}>
-        <DateRange
-          startDate={dateRangeStart}
-          endDate={dateRangeEnd}
-          onChangeEnd={onDateRangeEndChange}
-          onChangeStart={onDateRangeStartChange}
-          onSaveButton={onSaveDateRange}
-        />
-        {renderWeekDays()}
-      </div>
-    );
-  };
-
-  const adminConfigByDay = () => (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: '20px' }}>
-      <strong>Set as Weekend</strong>
-      <div
-        style={{
-          display: 'flex',
-          flexDirection: 'row',
-          justifyContent: 'space-between'
-        }}
-      >
-        {renderWeekendSettings()}
-      </div>
-      {renderConfigDays()}
-      <div>
-        <div>
-          <strong>Repeat Cycle</strong>
-        </div>
-        <Input
-          type="number"
-          name="cycleNumber"
-          placeholder="Please input number"
-        />
-        <Select
-          value={contentType}
-          onChange={onContentTypeSelect}
-          placeholder="How long"
-          options={['Days', 'Weeks', 'Months'].map(day => ({
-            value: day,
-            label: __(day)
-          }))}
-        />
-      </div>
-    </div>
-  );
-  const onSelectDateChange = dateString => {
-    if (dateString) {
-      const getDate = new Date(dateString).toDateString();
-
-      const newDates = scheduleDates;
-
-      newDates[getDate] = {
-        shiftStart: new Date(getDate),
-        shiftEnd: new Date(getDate)
-      };
-
-      setScheduleDates(newDates);
-      setKeyCounter(getDate);
-    }
-  };
-
-  const adminConfigBySelect = () => (
-    <>
-      <CustomRangeContainer>
-        <DateControl
-          required={false}
-          name="startDate"
-          onChange={onSelectDateChange}
-          placeholder={'Select date'}
-          dateFormat={'YYYY-MM-DD'}
-        />
-      </CustomRangeContainer>
-      {renderWeekDays()}
-    </>
-  );
-
-  const onContentTypeSelect = contntType => {
-    localStorage.setItem('contentType', JSON.stringify(contntType));
-    const contType = JSON.parse(localStorage.getItem('contentType') || '[]')
-      .value;
-    setContentType(contType);
-  };
-
-  const renderSwitchContent = () => {
-    switch (contentType) {
-      case 'By Date Selection':
-        return adminConfigBySelect();
-      case 'By Date Range':
-        return adminConfigByDate();
-      default:
-        return adminConfigByDay();
-    }
-  };
-
-  const removeSchedule = (_id: string, type: string) => {
-    removeScheduleShifts(_id, type);
-  };
-  const adminConfigDefaultContent = ({ closeModal }) => {
-    return (
-      <div style={{ display: 'flex', flexDirection: 'column', gap: '20px' }}>
-        <SelectTeamMembers
-          queryParams={queryParams}
-          label={'Team member'}
-          onSelect={onUserSelect}
-          name="userId"
-        />
-        <SelectDepartments
-          isRequired={false}
-          defaultValue={selectedDeptIds}
-          onChange={onDepartmentSelect}
-        />
-        <FormGroup>
-          <ControlLabel>Branches</ControlLabel>
-          <Row>
-            <Select
-              value={selectedBranchIds}
-              onChange={onBranchSelect}
-              placeholder="Select branch"
-              multi={true}
-              options={branchesList && renderBranchOptions(branchesList)}
-            />
-          </Row>
-        </FormGroup>
-        <Select
-          value={contentType}
-          onChange={onContentTypeSelect}
-          placeholder="Select Content Type"
-          options={['By Date Range', 'By Week Day', 'By Date Selection'].map(
-            day => ({
-              value: day,
-              label: __(day)
-            })
-          )}
-        />
-        {renderSwitchContent()}
-        <div
-          style={{
-            display: 'flex',
-            flexDirection: 'row',
-            justifyContent: 'center'
-          }}
-        >
-          <Button style={{ marginTop: 10 }} onClick={addDay}>
-            Add day
-          </Button>
-          <Button
-            style={{ marginTop: 10 }}
-            onClick={() => onAdminSubmitClick(closeModal)}
-          >
-            {'Submit'}
-          </Button>
-        </div>
-      </div>
     );
   };
 
@@ -619,31 +89,22 @@ function ScheduleList(props: Props) {
         size="lg"
         title={__('Schedule config - Admin')}
         trigger={adminConfigTrigger}
-        content={adminConfigDefaultContent}
+        content={adminConfigContent}
       />
     </>
   );
 
-  const title = (
-    <div
-      style={{
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        marginRight: '20px',
-        fontSize: '24px'
-      }}
-    />
-  );
-
   const actionBar = (
     <Wrapper.ActionBar
-      left={title}
       right={actionBarRight}
       hasFlex={true}
       wideSpacing={true}
     />
   );
+
+  const removeSchedule = (_id: string, type: string) => {
+    removeScheduleShifts(_id, type);
+  };
 
   const ListShiftContent = shifts => {
     return (
